+++ src/react-motion.js
@@ -1,0 +1,1 @@
+export {Spring, TransitionSpring, utils} from './Spring';